Added Logging Console Feature
MainContentPanel.mxml -> Added a vertical box divider to include
a textarea where we can have an auto scrolling text box that
displays, like a console, a log of all the user's inputs.

MainContentPanelMediator -> Added a datamember status for the console.
The console is just a string that grows and autoscrolls, therefore there
might be issues if the textarea grows too large.

Integrated a notification for the MainContentPanel that updates the
console when received.

Notifications -> Added a constant UPDATE STATUS notifications

MageBarMediator -> Added a wrapper function to send notifications
to update the status console.

Oligo -> Will use this for working with features

--- Future Work ----

The next step would be to use the post response variables to pass
an array of genbank files which can be used to load and edit a sequence