Added Plotting Functionality
Added working plotting functionality.
All three plots are displayed simulatneously at once and the resultant plotting is shown.

ApplicationFacade.as -> Added an update Charts function as a hack to get
the MainContentPanelMediator to be able to sync with the MageBar

MageBar.mxml -> integrated the updateCharts function into the oligo selection combox
changes

MageBarMediator -> Updated the onMerlinResonse Function to load the plot
data from the response

Oligo.as -> Added more getters/setters and fields for dealing with the chart
plotting data.

Notifications.as -> Added UPDATE_CHARTS notification

TODO: Still need to show the optMAGE vs Merlin points on the graph